requirements.txt
Django==2.2.16
# pytest==6.2.4
# pytest-django==4.4.0
# requests==2.26.0
django-debug-toolbar==3.2.4
